@@ -12,45 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.3",
-    "@angular/cdk": "^5.1.1",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "^2.0.0-beta.12",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.1.1",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@ngrx/store": "^5.1.0",
-    "angularfire2": "^5.0.0-rc.6",
-    "core-js": "^2.4.1",
-    "firebase": "^4.10.1",
+    "@angular/animations": "^6.0.4",
+    "@angular/cdk": "^6.2.1",
+    "@angular/common": "^6.0.4",
+    "@angular/compiler": "^6.0.4",
+    "@angular/core": "^6.0.4",
+    "@angular/flex-layout": "^6.0.0-beta.16",
+    "@angular/forms": "^6.0.4",
+    "@angular/http": "^6.0.4",
+    "@angular/material": "^6.2.1",
+    "@angular/platform-browser": "^6.0.4",
+    "@angular/platform-browser-dynamic": "^6.0.4",
+    "@angular/router": "^6.0.4",
+    "@ngrx/store": "^6.0.1",
+    "angularfire2": "^5.0.0-rc.10",
+    "core-js": "^2.5.7",
+    "firebase": "^5.0.4",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.5.6",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.2.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "^6.0.4",
+    "@angular/language-service": "^6.0.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~10.3.2",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~2.0.2",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~6.1.0",
+    "tslint": "~5.10.0",
+    "typescript": "~2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
