@@ -1,4 +1,4 @@
-import {AuthActions, SET_AUTHENTICATED, SET_UNAUTHENTICATED} from './auth.actions';
+import { AuthActions, SET_AUTHENTICATED, SET_UNAUTHENTICATED } from './auth.actions';
 
 export interface IState {
   isAuthenticated: boolean;
